backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
gunicorn
python-dotenv